refactor: Pagination 관련 interface 추가 및 관련 DTO들을 interface 확장으로 중복된 코드 제거
